move RatingSelector component in new file
